+++ QMA.Model/DeleteablePrimaryKeyBase.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace QMA.Model
+{
+    public abstract class DeleteablePrimaryKeyBase : PrimaryKeyBase
+    {
+        public DateTimeOffset? Deleted { get; set; }
+    }
+}