+++ QMA.Model/PrimaryKeyBase.cs
@@ -7,7 +7,5 @@ namespace QMA.Model
     public abstract class PrimaryKeyBase
     {
         public string PrimaryKey { get; set; }
-
-        public DateTimeOffset? Deleted { get; set; }
     }
 }